+++ app/src/androidTest/java/com/k0dm/atlacharacters/Character.kt
@@ -1,0 +1,8 @@
+package com.k0dm.atlacharacters
+
+internal class Character(
+    val name: String,
+    val allies: String,
+    val enemies: String,
+    val affiliation: String
+)
