+++ app/src/androidTest/java/com/k0dm/atlacharacters/ScenarioTest.kt
@@ -79,9 +79,3 @@ class ScenarioTest {
     }
 }
 
-internal class Character(
-    val name: String,
-    val allies: String,
-    val enemies: String,
-    val affiliation: String
-)
