+++ app/src/main/res/drawable/characters_icon.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0,0h24v24h-24z"/>
+    <path
+        android:pathData="M0,18V16.425C0,15.708 0.367,15.125 1.1,14.675C1.833,14.225 2.8,14 4,14C4.217,14 4.425,14.004 4.625,14.012C4.825,14.021 5.017,14.042 5.2,14.075C4.967,14.425 4.792,14.792 4.675,15.175C4.558,15.558 4.5,15.958 4.5,16.375V18H0ZM6,18V16.375C6,15.842 6.146,15.354 6.438,14.913C6.729,14.471 7.142,14.083 7.675,13.75C8.208,13.417 8.846,13.167 9.587,13C10.329,12.833 11.133,12.75 12,12.75C12.883,12.75 13.696,12.833 14.438,13C15.179,13.167 15.817,13.417 16.35,13.75C16.883,14.083 17.292,14.471 17.575,14.913C17.858,15.354 18,15.842 18,16.375V18H6ZM19.5,18V16.375C19.5,15.942 19.446,15.533 19.337,15.15C19.229,14.767 19.067,14.408 18.85,14.075C19.033,14.042 19.221,14.021 19.413,14.012C19.604,14.004 19.8,14 20,14C21.2,14 22.167,14.221 22.9,14.663C23.633,15.104 24,15.692 24,16.425V18H19.5ZM8.125,16H15.9C15.733,15.667 15.271,15.375 14.512,15.125C13.754,14.875 12.917,14.75 12,14.75C11.083,14.75 10.246,14.875 9.488,15.125C8.729,15.375 8.275,15.667 8.125,16ZM4,13C3.45,13 2.979,12.804 2.588,12.413C2.196,12.021 2,11.55 2,11C2,10.433 2.196,9.958 2.588,9.575C2.979,9.192 3.45,9 4,9C4.567,9 5.042,9.192 5.425,9.575C5.808,9.958 6,10.433 6,11C6,11.55 5.808,12.021 5.425,12.413C5.042,12.804 4.567,13 4,13ZM20,13C19.45,13 18.979,12.804 18.587,12.413C18.196,12.021 18,11.55 18,11C18,10.433 18.196,9.958 18.587,9.575C18.979,9.192 19.45,9 20,9C20.567,9 21.042,9.192 21.425,9.575C21.808,9.958 22,10.433 22,11C22,11.55 21.808,12.021 21.425,12.413C21.042,12.804 20.567,13 20,13ZM12,12C11.167,12 10.458,11.708 9.875,11.125C9.292,10.542 9,9.833 9,9C9,8.15 9.292,7.438 9.875,6.863C10.458,6.287 11.167,6 12,6C12.85,6 13.563,6.287 14.137,6.863C14.712,7.438 15,8.15 15,9C15,9.833 14.712,10.542 14.137,11.125C13.563,11.708 12.85,12 12,12ZM12,10C12.283,10 12.521,9.904 12.712,9.712C12.904,9.521 13,9.283 13,9C13,8.717 12.904,8.479 12.712,8.288C12.521,8.096 12.283,8 12,8C11.717,8 11.479,8.096 11.288,8.288C11.096,8.479 11,8.717 11,9C11,9.283 11.096,9.521 11.288,9.712C11.479,9.904 11.717,10 12,10Z"
+        android:fillColor="#000000"/>
+  </group>
+</vector>
