+++ app/src/main/res/drawable/favorite.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+  <group>
+    <clip-path
+        android:pathData="M2,2h44v44h-44z"/>
+    <path
+        android:pathData="M33.836,30.136C31.669,32.302 28.914,34.893 25.564,37.912C25.563,37.913 25.562,37.913 25.561,37.914L24,39.314L22.439,37.914C22.438,37.913 22.437,37.913 22.436,37.912C19.086,34.893 16.331,32.302 14.164,30.136C12.007,27.978 10.328,26.079 9.105,24.432C7.881,22.784 7.091,21.354 6.661,20.135C6.219,18.879 6,17.604 6,16.3C6,13.665 6.862,11.567 8.564,9.864C10.267,8.162 12.365,7.3 15,7.3C16.441,7.3 17.801,7.603 19.102,8.211C20.406,8.822 21.526,9.678 22.476,10.795L24,12.588L25.524,10.795C26.474,9.678 27.594,8.822 28.898,8.211C30.199,7.603 31.559,7.3 33,7.3C35.635,7.3 37.733,8.162 39.436,9.864C41.138,11.567 42,13.665 42,16.3C42,17.604 41.781,18.879 41.339,20.135C40.909,21.354 40.119,22.784 38.895,24.432C37.672,26.079 35.993,27.978 33.836,30.136Z"
+        android:strokeWidth="4"
+        android:fillColor="#F4A328"
+        android:strokeColor="#000000"/>
+  </group>
+</vector>
