+++ app/src/main/res/drawable/favorites_icon.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0,0h24v24h-24z"/>
+    <path
+        android:pathData="M12,21L10.55,19.7C8.867,18.183 7.475,16.875 6.375,15.775C5.275,14.675 4.4,13.688 3.75,12.813C3.1,11.938 2.646,11.133 2.388,10.4C2.129,9.667 2,8.917 2,8.15C2,6.583 2.525,5.275 3.575,4.225C4.625,3.175 5.933,2.65 7.5,2.65C8.367,2.65 9.192,2.833 9.975,3.2C10.758,3.567 11.433,4.083 12,4.75C12.567,4.083 13.242,3.567 14.025,3.2C14.808,2.833 15.633,2.65 16.5,2.65C18.067,2.65 19.375,3.175 20.425,4.225C21.475,5.275 22,6.583 22,8.15C22,8.917 21.871,9.667 21.612,10.4C21.354,11.133 20.9,11.938 20.25,12.813C19.6,13.688 18.725,14.675 17.625,15.775C16.525,16.875 15.133,18.183 13.45,19.7L12,21ZM12,18.3C13.6,16.867 14.917,15.637 15.95,14.613C16.983,13.587 17.8,12.696 18.4,11.938C19,11.179 19.417,10.504 19.65,9.912C19.883,9.321 20,8.733 20,8.15C20,7.15 19.667,6.317 19,5.65C18.333,4.983 17.5,4.65 16.5,4.65C15.717,4.65 14.992,4.871 14.325,5.312C13.658,5.754 13.2,6.317 12.95,7H11.05C10.8,6.317 10.342,5.754 9.675,5.312C9.008,4.871 8.283,4.65 7.5,4.65C6.5,4.65 5.667,4.983 5,5.65C4.333,6.317 4,7.15 4,8.15C4,8.733 4.117,9.321 4.35,9.912C4.583,10.504 5,11.179 5.6,11.938C6.2,12.696 7.017,13.587 8.05,14.613C9.083,15.637 10.4,16.867 12,18.3Z"
+        android:fillColor="#1C1B1F"/>
+  </group>
+</vector>
