+++ app/src/main/res/drawable/not_favorite.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+  <group>
+    <clip-path
+        android:pathData="M0,0h48v48h-48z"/>
+    <path
+        android:pathData="M24,42L21.1,39.4C17.733,36.367 14.95,33.75 12.75,31.55C10.55,29.35 8.8,27.375 7.5,25.625C6.2,23.875 5.292,22.267 4.775,20.8C4.258,19.333 4,17.833 4,16.3C4,13.167 5.05,10.55 7.15,8.45C9.25,6.35 11.867,5.3 15,5.3C16.733,5.3 18.383,5.667 19.95,6.4C21.517,7.133 22.867,8.167 24,9.5C25.133,8.167 26.483,7.133 28.05,6.4C29.617,5.667 31.267,5.3 33,5.3C36.133,5.3 38.75,6.35 40.85,8.45C42.95,10.55 44,13.167 44,16.3C44,17.833 43.742,19.333 43.225,20.8C42.708,22.267 41.8,23.875 40.5,25.625C39.2,27.375 37.45,29.35 35.25,31.55C33.05,33.75 30.267,36.367 26.9,39.4L24,42ZM24,36.6C27.2,33.733 29.833,31.275 31.9,29.225C33.967,27.175 35.6,25.392 36.8,23.875C38,22.358 38.833,21.008 39.3,19.825C39.767,18.642 40,17.467 40,16.3C40,14.3 39.333,12.633 38,11.3C36.667,9.967 35,9.3 33,9.3C31.433,9.3 29.983,9.742 28.65,10.625C27.317,11.508 26.4,12.633 25.9,14H22.1C21.6,12.633 20.683,11.508 19.35,10.625C18.017,9.742 16.567,9.3 15,9.3C13,9.3 11.333,9.967 10,11.3C8.667,12.633 8,14.3 8,16.3C8,17.467 8.233,18.642 8.7,19.825C9.167,21.008 10,22.358 11.2,23.875C12.4,25.392 14.033,27.175 16.1,29.225C18.167,31.275 20.8,33.733 24,36.6Z"
+        android:fillColor="#1C1B1F"/>
+  </group>
+</vector>
