+++ app/src/main/res/layout/character_info.xml
@@ -1,0 +1,97 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/contentLayout"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/nameTextView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:fontFamily="@font/poppins_bold"
+        android:textColor="@color/title_text_color"
+        android:textSize="32sp"
+        tools:text="Mako" />
+
+    <ImageView
+        android:id="@+id/characterImageView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:importantForAccessibility="no"
+        tools:src="@drawable/mako_image" />
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/background_category"
+        android:fontFamily="@font/poppins_bold"
+        android:text="@string/allies"
+        android:textAlignment="center"
+        android:textColor="@color/title_text_color"
+        android:textSize="24sp" />
+
+    <TextView
+        android:id="@+id/alliesTextView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/background_text"
+        android:fontFamily="@font/poppins_medium"
+        android:gravity="start|center"
+        android:maxLines="2"
+        android:text="@string/allies"
+        android:textColor="@color/black"
+        android:textSize="22sp"
+        tools:text="Bolin, Korra, Asami" />
+
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/background_category"
+        android:fontFamily="@font/poppins_bold"
+        android:text="@string/enemies"
+        android:textAlignment="center"
+        android:textColor="@color/title_text_color"
+        android:textSize="24sp" />
+
+    <TextView
+        android:id="@+id/enemiesTextView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/background_text"
+        android:fontFamily="@font/poppins_medium"
+        android:gravity="start|center"
+        android:maxLines="2"
+        android:text="@string/allies"
+        android:textColor="@color/black"
+        android:textSize="22sp"
+        tools:text="Hiroshi Sato, Amon, Tarrlok, Kuvira" />
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/background_category"
+        android:fontFamily="@font/poppins_bold"
+        android:text="@string/affiliation"
+        android:textAlignment="center"
+        android:textColor="@color/title_text_color"
+        android:textSize="24sp" />
+
+    <TextView
+        android:id="@+id/affiliationTextView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/background_text"
+        android:fontFamily="@font/poppins_medium"
+        android:gravity="start|center"
+        android:maxLines="3"
+        android:text="@string/allies"
+        android:textColor="@color/black"
+        android:textSize="22sp"
+        tools:text="Fire Ferrets (formerly) Republic City Police Republic City power plant (formerly)" />
+
+
+</LinearLayout>
