+++ app/src/main/res/layout/fragment_characters.xml
@@ -1,0 +1,118 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/characterLayout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/background"
+    android:orientation="vertical">
+
+    <LinearLayout
+        android:id="@+id/loadingLayout"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        android:orientation="vertical"
+        tools:visibility="gone">
+
+        <TextView
+            android:id="@+id/titleTextView"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:fontFamily="@font/poppins_bold"
+            android:text="@string/atla_characters"
+            android:textAlignment="center"
+            android:textColor="@color/black"
+            android:textSize="36sp"
+            tools:visibility="visible" />
+
+        <Space
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:layout_weight="1"
+            tools:visibility="visible" />
+
+        <TextView
+            android:id="@+id/errorTextView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginHorizontal="46dp"
+            android:fontFamily="@font/poppins_bold"
+            android:gravity="center"
+            android:textColor="@color/error"
+            android:textSize="24sp"
+            tools:text="Service unavailable"
+            tools:visibility="visible" />
+
+        <com.google.android.material.progressindicator.CircularProgressIndicator
+            style="@style/Widget.Material3.CircularProgressIndicator"
+            android:layout_width="64dp"
+            android:layout_height="64dp"
+            android:layout_gravity="center"
+            android:indeterminate="true"
+            app:indicatorColor="@color/secondary_container"
+            app:indicatorSize="64dp"
+            app:trackCornerRadius="10dp"
+            app:trackThickness="6dp"
+            tools:visibility="gone" />
+
+        <TextView
+            android:id="@+id/loadingDataTextView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:fontFamily="@font/poppins_regular"
+            android:text="@string/loading_character"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            tools:visibility="gone" />
+
+        <Space
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:layout_weight="1"
+            tools:visibility="visible" />
+
+    </LinearLayout>
+
+
+    <ImageView
+        android:id="@+id/favoriteImageView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end"
+        android:layout_margin="8dp"
+        android:importantForAccessibility="no"
+        android:src="@drawable/not_favorite" />
+
+    <LinearLayout
+        android:id="@+id/mainLayout"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_marginHorizontal="46dp"
+        android:layout_weight="1"
+        android:gravity="center">
+
+        <include layout="@layout/character_info" />
+
+    </LinearLayout>
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/actionButton"
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:layout_marginHorizontal="46dp"
+        android:layout_marginBottom="8dp"
+        android:backgroundTint="@color/background"
+        android:fontFamily="@font/poppins_medium"
+        android:paddingVertical="0dp"
+        android:textColor="@color/secondary_container"
+        android:textSize="24sp"
+        app:strokeColor="@color/black"
+        app:strokeWidth="2dp"
+        tools:text="Retry"
+        tools:visibility="visible" />
+
+</LinearLayout>
