+++ app/src/main/res/values/colors.xml
@@ -2,4 +2,11 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="surface_container">#FFFEDD9E</color>
+    <color name="secondary_container">#FFF4A328</color>
+    <color name="title_text_color">#FF715429</color>
+    <color name="background_category">#FFFFEED4</color>
+    <color name="background_text">#FFFFF8EE</color>
+    <color name="background">#FFFFFFEE</color>
+    <color name="error">#FF924242</color>
 </resources>
