+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,11 @@
 <resources>
     <string name="app_name">ATLA Characters</string>
+    <string name="characters_item">Characters</string>
+    <string name="favorites_item">Favorites</string>
+    <string name="allies">Allies</string>
+    <string name="enemies">Enemies</string>
+    <string name="affiliation">Affiliation</string>
+    <string name="loading_character">Loading character...</string>
+    <string name="atla_characters">ATLA Characters</string>
+    <string name="favorite_characters">Favorite characters</string>
 </resources>
