+++ app/src/main/res/values/themes.xml
@@ -5,5 +5,14 @@
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
+    <style name="Widget.App.BottomNavigationView" parent="Widget.Material3.BottomNavigationView">
+        <item name="materialThemeOverlay">@style/ThemeOverlay.App.BottomNavigationView</item>
+    </style>
+
+    <style name="ThemeOverlay.App.BottomNavigationView" parent="">
+        <item name="colorSecondaryContainer">@color/secondary_container</item>
+        <item name="colorSurfaceContainer">@color/surface_container</item>
+    </style>
+
     <style name="Theme.ATLACharacters" parent="Base.Theme.ATLACharacters" />
 </resources>
